Refactor components into separate files
- Get rid of loan repayment page
- Make sure ICR payment is atleast 0
